+++ src/test/java/org/apache/commons/math4/ode/nonstiff/DormandPrince853IntegratorTest.java
@@ -67,7 +67,7 @@ public class DormandPrince853IntegratorTest {
 
       double[] y0   = new double[k.length];
       for (int i = 0; i < y0.length; ++i) {
-          y0[i] = i + 1;
+          y0[i] = (double) i + 1;
       }
       double[] y    = new double[k.length];
